CharaSim: fix mob's HP&MP more than 2.1 billion

+++ WzComparerR2.Common/CharaSim/Mob.cs
@@ -26,8 +26,8 @@ namespace WzComparerR2.CharaSim
         public string DefaultMP { get; set; }
         public string FinalMaxHP { get; set; }
         public string FinalMaxMP { get; set; }
-        public int MaxHP { get; set; }
-        public int MaxMP { get; set; }
+        public long MaxHP { get; set; }
+        public long MaxMP { get; set; }
         public int HPRecovery { get; set; }
         public int MPRecovery { get; set; }
         public int? Speed { get; set; }
@@ -86,8 +86,8 @@ namespace WzComparerR2.CharaSim
                         case "defaultMP": mobInfo.DefaultMP = propNode.GetValueEx<string>(null); break;
                         case "finalmaxHP": mobInfo.FinalMaxHP = propNode.GetValueEx<string>(null); break;
                         case "finalmaxMP": mobInfo.FinalMaxMP = propNode.GetValueEx<string>(null); break;
-                        case "maxHP": mobInfo.MaxHP = propNode.GetValueEx<int>(0); break;
-                        case "maxMP": mobInfo.MaxMP = propNode.GetValueEx<int>(0); break;
+                        case "maxHP": mobInfo.MaxHP = propNode.GetValueEx<long>(0); break;
+                        case "maxMP": mobInfo.MaxMP = propNode.GetValueEx<long>(0); break;
                         case "hpRecovery": mobInfo.HPRecovery = propNode.GetValueEx<int>(0); break;
                         case "mpRecovery": mobInfo.MPRecovery = propNode.GetValueEx<int>(0); break;
                         case "speed": mobInfo.Speed = propNode.GetValueEx<int>(0); break;
